Moved filter functions to separate module
